statemachine_engine-0.0.9/requirements.txt
# Core dependencies
PyYAML>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
tabulate>=0.9.0
